seo_agent/requirements.txt
anthropic>=0.49.0
streamlit>=1.42.0
matplotlib>=3.9.0
japanize-matplotlib>=1.1.3
Pillow>=10.0.0
pydantic>=2.10.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
